added api return if token not present

+++ frontend/src/apis/task.js
@@ -3,15 +3,23 @@ import axios from "axios";
 const backendUrl = "http://localhost:3000/api/v1";
 
 export const createTask = async (task) => {
+    try{
         const reqUrl = `${backendUrl}/task/tasks`;
         const reqPayload = task;
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.post(reqUrl, reqPayload, {
             headers: {
                 'Authorization': `Bearer ${token}`
             }
         });
         return response.data;
+    }
+        catch (error) {
+            console.log(error);
+            return { success: false, message: "Cannot the create the task"}
+            // toast with custom message for clients
+        }
 };
 
 
@@ -19,6 +27,7 @@ export const getAllTasks = async () => {
     try {
         const reqUrl = `${backendUrl}/task/tasks`;
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.get(reqUrl, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -27,7 +36,7 @@ export const getAllTasks = async () => {
         return response.data;
     } catch (error) {
         console.log(error);
-        alert("Cannot the fetch the tasks")
+        return { success: false, message: "Cannot the fetch the tasks"}
         // toast with custom message for clients
     }
 };
@@ -39,7 +48,7 @@ export const getTask = async ({ id }) => {
         return response.data;
     } catch (error) {
         console.log(error);
-        alert("Cannot the fetch the task")
+        return { success: false, message: "Cannot the fetch the task"}
 
         // toast with custom message for clients
     }
@@ -50,6 +59,7 @@ export const updateTask = async (item) => {
         console.log(item)
         const reqUrl = `${backendUrl}/task/tasks/${item.id}`;
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.put(reqUrl,item.body, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -66,6 +76,7 @@ export const updateCompletedChecklist = async (item) => {
     try{
         const reqUrl = `${backendUrl}/task/tasks/${item.id}/checklist`;
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.put(reqUrl,item.checklist, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -84,6 +95,7 @@ export const analytics = async () => {
         const reqUrl = `${backendUrl}/task/tasks/analytics`;
         console.log({reqUrl})
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.get(reqUrl, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -93,7 +105,7 @@ export const analytics = async () => {
         return response.data;
     } catch (error) {
         console.log(error);
-        alert("Cannot the fetch the analytics")
+        return { success: false, message: "Cannot the fetch the analytics"}
 
         // toast with custom message for clients
     }
@@ -107,6 +119,7 @@ export const filterTask = async (filter) => {
         const reqPayload = { period:filter };
         console.log(reqPayload)
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.put(reqUrl, reqPayload, {
             headers: {
                 'Authorization': `Bearer ${token}`
@@ -116,19 +129,26 @@ export const filterTask = async (filter) => {
         return response.data;
     } catch (error) {
         console.log(error);
-        alert("Cannot the fetch the details")
+        return { success: false, message: "Cannot the fetch the details"}
 
         // toast with custom message for clients
     }
 };
 
 export const deleteTask = async ({id}) => {
+    try{
         const reqUrl = `${backendUrl}/task/tasks/${id}`;
         const token = localStorage.getItem('token'); // Replace 'token' with the key you used to store the token
+        if(!token)return;
         const response = await axios.delete(reqUrl, {
             headers: {
                 'Authorization': `Bearer ${token}`
             }
         });
         return response.data;
+    }catch (error) {
+        console.log(error);
+        return { success: false, message: "Cannot Delete"}
+        // toast with custom message for clients
+    }
 };
